CI: remove job [unit:lib-store-unix]
This job was created in !8328 to isolate of the store jobs, under
suspicion of flakiness. However, since then, this job has not been
flaky. Also, in !8291, when the store tests were ported to Alcotezt,
they started running with the other Tezts - consequently the store
tests now run twice. Consequently, removing this job.