MIR: Add a way to test-deploy rollup
To test everything works right we want to be able to deploy rollup.

This involves a ton of details that, I believe, should not be in the
mental context of the entire team, so this commit creates a Makefile
that performs everything on itself and adds instructions on how to use
it. What remains on the user is installing the necessary utilities,
providing an address with money, and running a couple of commands.

This deployment should be considered as test deployment, at least
because it does not allow configuring fee and burn caps.

The Makefile is pretty inspired by
https://github.com/baking-bad/tezos-on-tezos/, so I leave such the
license header.

I try to not use docker here, we in Serokell used to static binaries and
nix.